Move all the content to BeatmapSetHeaderContent drawable
